requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
streamlit>=1.20.0
PyPDF2>=3.0.0
python-docx>=0.8.11
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0 